chore(redux): rename logger
